+++ plugins/plugin-chart-composed/src/components/ComposedChart.tsx
@@ -103,6 +103,10 @@ const Styles = styled.div<ComposedChartStylesProps>`
   width: ${({ width }) => width}px;
   overflow: auto;
 
+  & .recharts-cartesian-axis-tick-line {
+    display: none;
+  }
+
   & .recharts-legend-item {
     cursor: pointer;
     white-space: nowrap;
@@ -155,7 +159,7 @@ const ComposedChart: FC<ComposedChartProps> = props => {
         setLegendWidth(currentWidth ? currentWidth + 20 : currentWidth);
       }
     }
-  }, [legendWidth, updater]);
+  }, [legendWidth, showLegend]);
 
   useEffect(() => {
     if (isSideLegend) {
@@ -183,6 +187,7 @@ const ComposedChart: FC<ComposedChartProps> = props => {
   const y2AxisHeight = Math.ceil(y2AxisClientRect?.height ?? 1);
   const y2AxisWidth = Math.ceil(y2AxisClientRect?.width ?? 1);
 
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   const updateVisibility = useCallback(
     debounce(() => {
       forceUpdate();
@@ -190,6 +195,8 @@ const ComposedChart: FC<ComposedChartProps> = props => {
     }, 5),
     [],
   );
+
+  // eslint-disable-next-line react-hooks/exhaustive-deps
   const updateUI = useCallback(
     debounce(() => {
       forceUpdate();
@@ -200,7 +207,7 @@ const ComposedChart: FC<ComposedChartProps> = props => {
 
   useEffect(() => {
     updateUI();
-  }, [props, forceUpdate]);
+  }, [props, forceUpdate, updateUI]);
 
   const currentData = useCurrentData(
     data,
@@ -230,7 +237,7 @@ const ComposedChart: FC<ComposedChartProps> = props => {
     legendPosition !== LegendPosition.left &&
     !yAxis.label
       ? xAxisHeight / 2 - yAxisWidth - 10
-      : 0;
+      : 5;
   const yMarginBottom =
     yAxis.tickLabelAngle === -45 && layout === Layout.vertical ? yAxisWidth - xAxisHeight - 10 : xAxisHeight;
 
@@ -246,8 +253,8 @@ const ComposedChart: FC<ComposedChartProps> = props => {
         layout={layout}
         style={{ visibility: visible ? 'visible' : 'hidden' }}
         margin={{
-          right: layout === Layout.vertical ? 10 : 0,
-          left: xMarginLeft > 0 ? xMarginLeft : 0,
+          right: layout === Layout.vertical ? 10 : 5,
+          left: xMarginLeft > 0 ? xMarginLeft : 5,
           top: 15,
           bottom: showLegend && legendPosition === LegendPosition.bottom ? 0 : yMarginBottom,
         }}