+++ plugins/plugin-chart-waterfall/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.4](https://github.com/nielsen-oss/superset-viz-plugins/compare/@superset-viz-plugins/plugin-chart-waterfall@0.10.3...@superset-viz-plugins/plugin-chart-waterfall@0.10.4) (2021-05-09)
+
+**Note:** Version bump only for package @superset-viz-plugins/plugin-chart-waterfall
+
+
+
+
+
+## 0.10.3 (2021-05-06)
+
+**Note:** Version bump only for package @superset-viz-plugins/plugin-chart-waterfall
+
+
+
+
+
 ## [0.10.2](https://github.com/nielsen-oss/superset-viz-plugins/compare/@superset-viz-plugins/plugin-chart-waterfall@0.10.1...@superset-viz-plugins/plugin-chart-waterfall@0.10.2) (2021-05-03)
 
 **Note:** Version bump only for package @superset-viz-plugins/plugin-chart-waterfall